Move printer files to a printer mod
