+++ src/lib.rs
@@ -4,6 +4,7 @@ pub use error::{LeetUpError, Result};
 pub mod cmd;
 mod config;
 mod error;
+mod printer;
 
 pub(crate) mod client;
 pub(crate) mod icon;
