+++ src/printer/mod.rs
@@ -1,0 +1,6 @@
+mod printer;
+mod test_execution_printer;
+
+pub use printer::*;
+pub use test_execution_printer::TestExecutionResult;
+
